requirements.txt
pandas
scikit-surprise